factorial done

+++ 0x08-recursion/3-factorial.c
@@ -5,12 +5,14 @@
  * returns factorial
  */
 
-int factorial(int n){
-	if (n == 1) {
+unsigned long long int factorial(unsigned int n)
+{
+	if (n == 1)
+	{
 		return (1);
-	} 
-	else {
-		return (n*factorial(n - 1));
-				}
-				}
-
+	}
+	else
+	{
+		return (n * factorial(n - 1));
+	}
+}
